# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DevMind AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SecurityScanInput {
  /** Source code to analyze */
  code: string;
  /** Programming language (e.g. python, javascript, java) */
  language: string;
  /**
   * Optional filename for context
   * @nullable
   */
  filename?: string | null;
}

export interface SecurityScan {
  id: number;
  code: string;
  language: string;
  /** @nullable */
  filename?: string | null;
  /** pending | complete | error */
  status: string;
  /**
   * 0-10 severity score
   * @nullable
   */
  severityScore?: number | null;
  /**
   * JSON array of vulnerability findings
   * @nullable
   */
  vulnerabilities?: string | null;
  /**
   * AI-generated summary
   * @nullable
   */
  summary?: string | null;
  createdAt: string;
}

export interface SecurityStats {
  totalScans: number;
  criticalCount: number;
  highCount: number;
  mediumCount: number;
  lowCount: number;
  avgSeverity: number;
}

export interface WorkflowInput {
  /** Natural language description of the automation */
  description: string;
}

export interface Workflow {
  id: number;
  description: string;
  /** pending | complete | error */
  status: string;
  /**
   * JSON representing the pipeline steps
   * @nullable
   */
  pipelineJson?: string | null;
  /**
   * JSON representing the visual flow diagram nodes/edges
   * @nullable
   */
  diagramJson?: string | null;
  createdAt: string;
}

export interface CodebaseProjectInput {
  name: string;
  /** @nullable */
  description?: string | null;
  /**
   * JSON string of file paths and contents
   * @nullable
   */
  files?: string | null;
}

export interface CodebaseProject {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** pending | indexed | error */
  status: string;
  /** @nullable */
  fileCount?: number | null;
  /**
   * JSON of indexed file contents
   * @nullable
   */
  files?: string | null;
  createdAt: string;
}

export interface CodebaseQueryInput {
  question: string;
}

export interface CodebaseQueryResult {
  id: number;
  projectId: number;
  question: string;
  answer: string;
  /**
   * JSON array of source references (file, line)
   * @nullable
   */
  sources?: string | null;
  createdAt: string;
}

export interface IncidentInput {
  /** Error logs, stack traces, or system metrics to analyze */
  logInput: string;
}

export interface Incident {
  id: number;
  logInput: string;
  /** pending | complete | error */
  status: string;
  /** @nullable */
  rootCause?: string | null;
  /** @nullable */
  affectedComponent?: string | null;
  /**
   * 0-1 confidence score
   * @nullable
   */
  confidence?: number | null;
  /**
   * Step-by-step fix instructions
   * @nullable
   */
  remediation?: string | null;
  /**
   * critical | high | medium | low
   * @nullable
   */
  severity?: string | null;
  createdAt: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface DashboardStats {
  totalScans: number;
  totalWorkflows: number;
  totalProjects: number;
  totalIncidents: number;
  /** Scans in the last 7 days */
  recentScans: number;
  avgSeverityScore: number;
}

export interface ActivityItem {
  id: number;
  /** security | workflow | codebase | analyzer */
  module: string;
  title: string;
  status: string;
  /** @nullable */
  severity?: string | null;
  createdAt: string;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
};