# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DevMind AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /codebase/projects:
    get:
      operationId: listCodebaseProjects
      tags: [codebase]
      summary: List all indexed codebase projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CodebaseProject"
    post:
      operationId: createCodebaseProject
      tags: [codebase]
      summary: Index a new codebase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodebaseProjectInput"
      responses:
        "201":
          description: Project indexed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CodebaseProject"
  /codebase/projects/{id}/query:
    post:
      operationId: queryCodebase
      tags: [codebase]
      summary: Ask a question about an indexed codebase
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodebaseQueryInput"
      responses:
        "200":
          description: Query answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CodebaseQueryResult"
  /codebase/projects/{id}/questions:
    get:
      operationId: listCodebaseQuestions
      tags: [codebase]
      summary: List past questions for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Question history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CodebaseQueryResult"

  # Root Cause Analyzer
  /analyzer/incidents:
    get:
      operationId: listIncidents
      tags: [analyzer]
      summary: List all analyzed incidents
      responses:
        "200":
          description: List of incidents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Incident"
    post:
      operationId: createIncident
      tags: [analyzer]
      summary: Analyze logs or error traces for root cause
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidentInput"
      responses:
        "201":
          description: Incident analyzed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
  /analyzer/incidents/{id}:
    get:
      operationId: getIncident
      tags: [analyzer]
      summary: Get a specific incident analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Incident details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
        "404":
          description: Not found

  # Dashboard / History
...
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to returnTo on success or /api/login on failure.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to OIDC provider logout URL.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Token did not contain usable claims.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Token exchange failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    CodebaseProjectInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        files:
          type: ["string", "null"]
          description: JSON string of file paths and contents
    CodebaseProject:
      type: object
      required: [id, name, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          description: "pending | indexed | error"
        fileCount:
          type: ["integer", "null"]
        files:
          type: ["string", "null"]
          description: JSON of indexed file contents
        createdAt:
          type: string
          format: date-time
    CodebaseQueryInput:
      type: object
      required: [question]
      properties:
        question:
          type: string
    CodebaseQueryResult:
      type: object
      required: [id, projectId, question, answer, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        question:
          type: string
        answer:
          type: string
        sources:
          type: ["string", "null"]
          description: JSON array of source references (file, line)
        createdAt:
          type: string
          format: date-time

    # Analyzer
    IncidentInput:
      type: object
      required: [logInput]
      properties:
        logInput:
          type: string
          description: Error logs, stack traces, or system metrics to analyze
    Incident:
      type: object
      required: [id, logInput, status, createdAt]
      properties:
        id:
          type: integer
        logInput:
          type: string
        status:
          type: string
          description: "pending | complete | error"
        rootCause:
          type: ["string", "null"]
        affectedComponent:
          type: ["string", "null"]
        confidence:
          type: ["number", "null"]
          description: 0-1 confidence score
        remediation:
          type: ["string", "null"]
          description: Step-by-step fix instructions
        severity:
          type: ["string", "null"]
          description: "critical | high | medium | low"
        createdAt:
          type: string
          format: date-time

    # Auth
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: ["string", "null"]
          format: email
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        profileImageUrl:
          type: ["string", "null"]
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: "null"
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # Dashboard
...